Change JS String.replaceAll() to regexp replace /.../gi for better client compat; Change psql deductions.id to bigint

+++ install/pg/create_tables.sql
@@ -93,7 +93,7 @@ CREATE TABLE adjustments (
 );
 
 CREATE TABLE deductions (
-	id integer primary key,
+	id bigint primary key,
 	routine_id integer references routines (id) on delete cascade,
 	judge_tag varchar,
 	ts bigint,
